// src/api/model/address.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * 互帮Api
 * OpenAPI spec version: v1
 */

/**
 * 地址
 */
export interface Address {
  id?: number;
  userId?: number;
  /**
   * 街道/小区
   * @nullable
   */
  street?: string | null;
  /**
   * 楼号/门牌号
   * @nullable
   */
  unit?: string | null;
  /**
   * 联系人
   * @nullable
   */
  contact?: string | null;
  /**
   * 手机
   * @nullable
   */
  phone?: string | null;
  /** 是否设置为默认 */
  isDefault?: boolean;
}